Make install script do the right things.
Previously it would do some wonky shit with the directories that were
being linked. This is now no longer a problem, as there is now actually
logic inside of the symlink loop.

Also check that zsh is installed before trying to change the shell, and
require that to be installed before moving on to anything else.